+++ src/ClientBridge.lua
@@ -140,6 +140,7 @@ function ClientBridge._start(config)
 				remote = remote,
 				args = params,
 			})
+			print(ReceiveQueue)
 		end
 	end)
 end