+++ tests/client/ClientTest2.client.lua
@@ -5,19 +5,37 @@ local BridgeNet = require(ReplicatedStorage.Packages.BridgeNet)
 BridgeNet.Start({})
 
 local Bridges = BridgeNet.CreateBridgeTree({
-	RemoteA = BridgeNet.Bridge({}),
+	RemoteA = BridgeNet.Bridge({
+		ReplicationRate = 20,
+	}),
 	RemoteCategory = {
-		RemoteB = BridgeNet.Bridge({
-			Name = "RemoteB",
-		}),
-		RemoteC = BridgeNet.Bridge({
-			Name = "RemoteC",
-		}),
+		RemoteB = BridgeNet.Bridge({}),
+		RemoteC = BridgeNet.Bridge({}),
 	},
 })
 
+if Bridges["RemoteA"] == nil then
+	error(".Declare error, RemoteA is nil")
+end
+
+if Bridges["RemoteCategory"] == nil then
+	error("RemoteCategory is nil")
+end
+
+if Bridges.RemoteCategory["RemoteB"] == nil then
+	error(".Declare error, RemoteB is nil")
+end
+
+if Bridges.RemoteA["_id"] == nil and Bridges.RemoteA["_name"] == nil then
+	error(".Declare does not return a bridge")
+end
+
 Bridges.RemoteA:Connect(function(arg1)
-	print("received:")
 	print(arg1)
-	print("--------------")
 end)
+
+while true do
+	Bridges.RemoteCategory.RemoteB:Fire("client to server check")
+	Bridges.RemoteCategory.RemoteB:InvokeServer("Args working")
+	task.wait(2)
+end